plotter-py/requirements.txt
cassandra-driver
pandas
matplotlib
